# openapi.yaml
openapi: 3.0.0
info:
  title: 'Speech Synthesis'
  description: 'Speech Synthesis API'
  license:
    name: MIT
    url: 'https://mit-license.org/2023/license.txt'
  version: 2026.04.2
servers:
  -
    url: 'http://localhost'
paths:
  /api/translate:
    post:
      tags:
        - 'Core Components'
      description: 'Translate a message key or text using the application locale catalog.'
      operationId: 'Translate text'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - text
              properties:
                text:
                  description: 'Message key or text to translate'
                  type: string
                  nullable: false
                lang:
                  description: 'Target locale (e.g. en, fr). Uses the request/app locale when omitted.'
                  type: string
                  example: fr
                  nullable: true
              type: object
      responses:
        '200':
          description: 'Translated message'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      description: 'This is the default error response for the API.'
      type: object
      allOf:
        -
          $ref: '#/components/schemas/SuccessResponse'
        -
          properties:
            success:
              description: 'Response status'
              type: boolean
              example: false
              nullable: false
          type: object
...
    SuccessResponse:
      description: 'This is the default response for the API.'
      properties:
        success:
          description: 'Response status'
          type: boolean
          nullable: false
        message:
          description: 'Status/Error message'
          type: string
          nullable: true
      type: object
...